Don't need a cast, the calling type is appropriate.
--HG--
extra : rebase_source : 2091fe521875de4a1b1f609e2189e99a114c8e1f

+++ src/os_posix/os_alloc.c
@@ -34,7 +34,7 @@ __wt_calloc(WT_SESSION_IMPL *session, size_t number, size_t size, void *retp)
 	if (session != NULL && S2C(session)->stats != NULL)
 		WT_CSTAT_INCR(session, memalloc);
 
-	if ((p = calloc(number, (size_t)size)) == NULL) {
+	if ((p = calloc(number, size)) == NULL) {
 		__wt_err(session, errno, "memory allocation");
 		return (WT_ERROR);
 	}
